+++ simbad/command_line/simbad_database.py
@@ -29,7 +29,7 @@ import simbad.db
 import simbad.exit
 import simbad.rotsearch.amore_search
 
-from simbad.util.pdb_util import get_pdb_content
+from simbad.util.pdb_util import PdbStructure
 
 logger = None
 
@@ -279,7 +279,7 @@ def create_contaminant_db(database, add_morda_domains, nproc=2, submit_qtype=Non
             if not os.path.exists(stem):
                 os.makedirs(stem)
 
-            content = get_pdb_content(result.pdb_code)
+            content = PdbStructure.get_pdb_content(result.pdb_code)
             if content is None:
                 logger.debug("Encountered a problem downloading PDB %s - skipping entry", result.pdb_code)
             else:
@@ -311,7 +311,7 @@ def create_contaminant_db(database, add_morda_domains, nproc=2, submit_qtype=Non
         if len(what_to_do) > 0:
             scripts, _, tmps, files = zip(*what_to_do)
             j = Job(submit_qtype)
-            j.submit(scripts, name='cont_db', nproc=nproc, submit_queue=submit_queue)
+            j.submit(scripts, name='cont_db', nproc=nproc, queue=submit_queue)
             j.wait()
 
             for output, final in files:
@@ -422,7 +422,7 @@ def create_morda_db(database, nproc=2, submit_qtype=None, submit_queue=False, ch
         # Run the scripts
         scripts, _, tmps, files = zip(*what_to_do)
         j = Job(submit_qtype)
-        j.submit(scripts, name='morda_db', nproc=nproc, submit_queue=submit_queue)
+        j.submit(scripts, name='morda_db', nproc=nproc, queue=submit_queue)
         j.wait()
 
         sub_dir_names = set([os.path.basename(f).rsplit('.', 1)[0][1:3] for f in chunk_dat_files])