+++ simbad/rotsearch/amore_score.py
@@ -36,6 +36,31 @@ class AmoreRotationScore(object):
         return string.format(name=self.__class__.__name__, **{k: getattr(self, k) for k in self.__slots__})
 
     def _as_dict(self):
-        """Convert the :obj:`_AmoreRotationScore <simbad.rotsearch.amore_search._AmoreRotationScore>`
+        """Convert the :obj:`AmoreRotationScore <simbad.rotsearch.amore_score.AmoreRotationScore>`
+        object to a dictionary"""
+        return {k: getattr(self, k) for k in self.__slots__}
+
+
+class DatModelScore(object):
+    """A dat model scoring class"""
+
+    __slots__ = ("pdb_code", "dat_path", "mw_diff", "x", "y", "z", "intrad")
+
+    def __init__(self, pdb_code, dat_path,  mw_diff, x, y, z, intrad):
+        self.pdb_code = pdb_code
+        self.dat_path = dat_path
+        self.mw_diff = mw_diff
+        self.x = x
+        self.y = y
+        self.z = z
+        self.intrad = intrad
+
+    def __repr__(self):
+        string = "{name}(pdb_code={pdb_code} dat_path={dat_path} " \
+                 "mw_diff={mw_diff} x={x} y={y} z={z} intrad={intrad})"
+        return string.format(name=self.__class__.__name__, **{k: getattr(self, k) for k in self.__slots__})
+
+    def _as_dict(self):
+        """Convert the :obj:`DatModelScore <simbad.rotsearch.amore_score.DatModelScore>`
         object to a dictionary"""
         return {k: getattr(self, k) for k in self.__slots__}